pyproject.toml
[project]
name = "github-intelligence-env"
version = "0.1.0"
description = """
A Python package for GitHub Intelligence.
"""
requires-python = ">=3.12,<3.13"

dependencies = [
    # "cached-path>=1.7.3",
    # "detect-secrets>=1.5.0",
    # "detoxify>=0.5.2",
    "fastapi>=0.115.13",
    "fastembed>=0.7.1",
    # "fastembed-gpu>=0.7.1",
    "guardrails-ai>=0.5.15",
    "guardrails-api-client>=0.3.13,<0.4.0",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pydantic-settings>=2.10.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.3",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
    "google-cloud-secret-manager>=2.21.1",
    "detect-secrets>=1.5.0",
    "cached-path>=1.7.3",
    "torch>=2.7.0",
    "detoxify>=0.5.2",
    "nltk>=3.8.1",
    "supabase>=2.27.0",
    "alembic>=1.17.2",
    # OpenTelemetry for observability
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-requests>=0.42b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.42b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.21.0",
    "opentelemetry-exporter-prometheus>=0.42b0",
    "prometheus-client>=0.19.0",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

# [[tool.uv.index]]
# name = "pytorch-cu126"
# url = "https://download.pytorch.org/whl/cu126"
# explicit = true


[dependency-groups]
dev = [
 "langgraph-cli[inmem]>=0.3.3",
 "pre-commit>=4.2.0",
]
lint = [
    "mypy>=1.16.1",
    "ruff>=0.12.0",
    "types-pyyaml>=6.0.12.20250516",
    "types-requests>=2.32.4.20250611",
]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

######################################
# --- Linting & Formatting Tools --- #
######################################

[tool.ruff]
# Assume Python 3.12
target-version = "py312"

# Same as Black.
line-length = 125
indent-width = 4

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

# Whether to show an enumeration of all fixed lint violations
show-fixes = true

# Enable common lint rules.
lint.select = [
  "B",    # flake8-bugbear
  "E",    # pycodestyle
  "F",    # Pyflakes1
  "I",    # isort
  "SIM",  # similarity
  "UP",   # pyupgrade
]
lint.ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]
lint.unfixable = []

[tool.ruff.lint.mccabe]
# Maximum allowed McCabe complexity.
max-complexity = 10


#########################
# --- Static Typing --- #
#########################

[tool.mypy]
# Mypy configuration.
files = ["src"]  # Check from project root instead of just src
mypy_path = ["src"]  # Set mypy path to project root
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_return_any =false
strict_optional = true

#########################
# --- Testing Tools --- #
#########################

[tool.pytest.ini_options]
testpaths = [ "tests" ]
python_files = [ "test_*.py" ]
addopts = "-ra -v -s"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
]
